refactor: move mcp_server.py into mcp/ (shim at old path)
